Start to implementation of cossym in perl
For portability this is a perl implementation of the original
shell scripts with unix utils 'cossym' program.  The main
remaining requirement for COS is the GNU make based build system.
At the moment, I plan to postpone that issue until the general
implementation of COS is working in unix/mac/windows.